feat:  Multiple Datasources Test for the RDBMS module using test templates #21983
- Moved RDBMS ITests to global QA Folder
- Fix Liquibase script for timestamp in different database

+++ db/rdbms/src/main/java/io/camunda/db/rdbms/sql/ProcessInstanceMapper.java
@@ -18,7 +18,7 @@ public interface ProcessInstanceMapper {
 
   ProcessInstanceEntity findOne(Long processInstanceKey);
 
-  Integer count(ProcessInstanceDbQuery filter);
+  Long count(ProcessInstanceDbQuery filter);
 
   List<ProcessInstanceEntity> search(ProcessInstanceDbQuery filter);
 }
